라우트 보호 기능 추가 및 유저 디테일 페이지에 비디오 추가

+++ controllers/videoControllers.js
@@ -24,8 +24,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
 
   res.redirect(routes.videoDetail(newVideo.id));
 };
@@ -36,7 +39,8 @@ export const videoDetail = async (req, res) => {
   } = req;
 
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator"); //mongoose.Schema.Types.ObjectId여기에다가만 쓸 수 있음
+    console.log(video);
 
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (err) {
@@ -52,7 +56,11 @@ export const getEditVideo = async (req, res) => {
 
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: "Edit Video", video });
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: "Edit Video", video });
+    }
   } catch (err) {
     res.redirect(routes.home);
   }
@@ -78,7 +86,12 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (err) {
     console.log(err);
   }
